Decentralized_PLN
The experiment result is shown in AccuracyTest.xlsx

The result is good for ONLY 2 NODES now, and when I use 5 nodes as paper does, the accuracy is around 50% (compared with 56% in the paper).

The probable parameter could be the eps0 (when we optimize O matrix in the ADMM.py) and the convergence condition (in the main function).